docs(abstractions): Fix XML documentation filename in `IHeaderedEnvelope{out TValue}.cs`
This commit corrects the `file` attribute in the XML documentation header of `src/IHeaderedEnvelope{out TValue}.cs`. The previous value `IHeaderedEnvelope2.cs` was a remnant of an old naming convention. It has been updated to `IHeaderedEnvelope{out TValue}.cs` to accurately reflect the current file name.

This is a documentation-only change that improves the accuracy and consistency of the codebase's XML comments.